Move MessageStack to kas::runner
